feat: customer helper 함수 분리
