+++ src/customer-profile/customer-profile.service.ts
@@ -16,6 +16,7 @@ export class CustomerProfileService {
   constructor(
     @InjectRepository(CustomerProfile)
     private readonly customerProfileRepository: Repository<CustomerProfile>,
+
     private readonly userService: UserService,
   ) {}
 
@@ -91,20 +92,5 @@ export class CustomerProfileService {
     return {
       message: '고객님의 프로필이 성공적으로 업데이트되었습니다.',
     };
-  }
-
-  public async getCustomerId(userId: string) {
-    const customer = await this.customerProfileRepository.findOne({
-      where: { user: { id: userId } },
-      select: ['id'],
-    });
-
-    if (!customer) {
-      throw new NotFoundException(
-        '고객님의 프로필을 찾을 수 없습니다, 프로필을 생성해주세요!',
-      );
-    }
-
-    return customer.id;
   }
 }
